Decompose Logger, Target, Console, Formatter from metalog.js
